+++ test/unit/test/protocol/null_bitmap_traits.cpp
@@ -5,7 +5,7 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#include <boost/mysql/impl/internal/protocol/null_bitmap_traits.hpp>
+#include <boost/mysql/impl/internal/protocol/impl/null_bitmap_traits.hpp>
 
 #include <boost/test/data/monomorphic/collection.hpp>
 #include <boost/test/data/test_case.hpp>
