+++ test/unit/test/protocol/protocol_field_type.cpp
@@ -8,7 +8,7 @@
 #include <boost/mysql/column_type.hpp>
 #include <boost/mysql/mysql_collations.hpp>
 
-#include <boost/mysql/impl/internal/protocol/protocol_field_type.hpp>
+#include <boost/mysql/impl/internal/protocol/impl/protocol_field_type.hpp>
 
 #include <boost/test/unit_test.hpp>
 
@@ -19,6 +19,49 @@ using namespace boost::mysql::detail;
 using namespace boost::mysql::test;
 namespace collations = boost::mysql::mysql_collations;
 using boost::mysql::column_type;
+
+namespace boost {
+namespace mysql {
+namespace detail {
+
+// protocol_field_type
+inline std::ostream& operator<<(std::ostream& os, protocol_field_type t)
+{
+    switch (t)
+    {
+    case detail::protocol_field_type::decimal: return os << "decimal";
+    case detail::protocol_field_type::tiny: return os << "tiny";
+    case detail::protocol_field_type::short_: return os << "short_";
+    case detail::protocol_field_type::long_: return os << "long_";
+    case detail::protocol_field_type::float_: return os << "float_";
+    case detail::protocol_field_type::double_: return os << "double_";
+    case detail::protocol_field_type::null: return os << "null";
+    case detail::protocol_field_type::timestamp: return os << "timestamp";
+    case detail::protocol_field_type::longlong: return os << "longlong";
+    case detail::protocol_field_type::int24: return os << "int24";
+    case detail::protocol_field_type::date: return os << "date";
+    case detail::protocol_field_type::time: return os << "time";
+    case detail::protocol_field_type::datetime: return os << "datetime";
+    case detail::protocol_field_type::year: return os << "year";
+    case detail::protocol_field_type::varchar: return os << "varchar";
+    case detail::protocol_field_type::bit: return os << "bit";
+    case detail::protocol_field_type::newdecimal: return os << "newdecimal";
+    case detail::protocol_field_type::enum_: return os << "enum_";
+    case detail::protocol_field_type::set: return os << "set";
+    case detail::protocol_field_type::tiny_blob: return os << "tiny_blob";
+    case detail::protocol_field_type::medium_blob: return os << "medium_blob";
+    case detail::protocol_field_type::long_blob: return os << "long_blob";
+    case detail::protocol_field_type::blob: return os << "blob";
+    case detail::protocol_field_type::var_string: return os << "var_string";
+    case detail::protocol_field_type::string: return os << "string";
+    case detail::protocol_field_type::geometry: return os << "geometry";
+    default: return os << "unknown";
+    }
+}
+
+}  // namespace detail
+}  // namespace mysql
+}  // namespace boost
 
 BOOST_AUTO_TEST_SUITE(test_protocol_field_type)
 
@@ -38,10 +81,8 @@ BOOST_AUTO_TEST_CASE(compute_column_type_legacy_types)
     } test_cases[] = {
         {"tiny_text",      protocol_field_type::tiny_blob,   0, collations::utf8mb4_general_ci, column_type::text     },
         {"tiny_blob",      protocol_field_type::tiny_blob,   0, collations::binary,             column_type::blob     },
-        {"medium_text",
-         protocol_field_type::medium_blob,
-         0,                                                     collations::utf8mb4_general_ci,
-         column_type::text                                                                                            },
+        {"medium_text",    protocol_field_type::medium_blob, 0, collations::utf8mb4_general_ci, column_type::text
+        },
         {"medium_blob",    protocol_field_type::medium_blob, 0, collations::binary,             column_type::blob     },
         {"long_text",      protocol_field_type::long_blob,   0, collations::utf8mb4_general_ci, column_type::text     },
         {"long_blob",      protocol_field_type::long_blob,   0, collations::binary,             column_type::blob     },
