+++ test/unit/test/protocol/serialization_context.cpp
@@ -1,0 +1,18 @@
+//
+// Copyright (c) 2019-2024 Ruben Perez Hidalgo (rubenperez038 at gmail dot com)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
+#include <boost/mysql/impl/internal/protocol/impl/serialization_context.hpp>
+
+#include <boost/test/unit_test.hpp>
+
+using namespace boost::mysql;
+
+BOOST_AUTO_TEST_SUITE(test_serialization_context)
+
+// TODO
+
+BOOST_AUTO_TEST_SUITE_END()
